ai-service/pyproject.toml
package-mode = false

[tool.poetry]
name = "ai-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.8"
langgraph = "^0.2.53"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
langchain-openai = "^0.2.9"
langchain-core = "^0.3.21"
python-dotenv = "^1.0.1"
langgraph-checkpoint-postgres = "^2.0.3"
psycopg-binary = "^3.2.3"
langchain-postgres = "^0.0.12"
langchain-community = "^0.3.8"
beautifulsoup4 = "^4.12.3"
python-multipart = "^0.0.17"
wikipedia = "^1.4.0"
pdfminer-six = "^20240706"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
greenlet = "^3.1.1"
pydantic = {extras = ["email"], version = "^2.10.2"}
langchain-google-community = "^2.0.3"
google-api-python-client = "^2.154.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
sse-starlette = "^2.1.3"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.32.1"


[tool.poetry.group.lint.dependencies]
ruff = "^0.8.0"
codespell = "^2.3.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
hypothesis = "^6.119.4"
tox = "^4.23.2"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
